Example bldg 1
Added diagrams and partial code for the building example.

Still need to implement all properties and datatypes (all for energy
and some for location).

Still need to remove the properties and datatypes not used. Don’t know
what to do about the hierarchies, whether to keep only the last class
or the hierarchy (so far the hierarchy is keep only for the branch used)